@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import {HttpClient} from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import {BehaviorSubject, firstValueFrom, map, Observable, tap} from "rxjs";
 import {createEvent, Event} from "../models/event.model";
 import {environment} from "../../../environments/environment";
