@@ -1,4 +1,4 @@
-import {HttpClient} from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {DirectoryItemsStore} from './directory-items.store';
 import {firstValueFrom, tap} from 'rxjs';
