@@ -1,4 +1,4 @@
-import {HttpClient} from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {GoalsStore, GoalsUiState} from './goals.store';
 import {firstValueFrom, tap} from "rxjs";
