@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import {HttpClient} from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import {BehaviorSubject, firstValueFrom, Observable, tap} from "rxjs";
 import {environment} from "../../../environments/environment";
 
